examples/readitlater/ogp/requirements.txt
python-opengraph